requirements.txt
asgiref==3.3.1
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.2
djongo==1.3.4
mongoengine==0.22.1
pymongo==3.11.3
pytz==2021.1
sqlparse==0.2.4
